fix: remember unparsed bytes when needing more data
The function `enqueue_messages_from_buffer` is responsible for splitting
received bytes from bird into messages. if a message is not entirely
contained within the data in the buffer, new data has to be read. this
happens outside of this function. before returning, the function has
to make sure the current buffer is remembered. this commit remembers
the unparsed data in the case where a single message is larger than
the buffer. this is done by not returning early, allowing the later part
of the function to remember the unparsed data.